fix norminette

+++ minitalk/server.c
@@ -6,7 +6,7 @@
 /*   By: hkubo <hkubo@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/01 18:27:23 by hkubo             #+#    #+#             */
-/*   Updated: 2021/08/08 19:43:07 by hkubo            ###   ########.fr       */
+/*   Updated: 2021/08/08 19:43:15 by hkubo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -89,7 +89,7 @@ int	main(int argc, char *argv[])
 
 	(void)argc;
 	(void)argv;
-	pid = ft_itoa(getpid());
+	pid = ft_itoa((int)getpid());
 	printf("len: %d\n", (int)ft_strlen(pid));
 	ft_bzero(&sigusr, sizeof(struct sigaction));
 	sigusr.sa_sigaction = sigusr_handler;
